Refactoring required for blueprint xenapi-live-migration
This refactoring of the libvirt live migration code is
required to enable live migration in the xenapi driver.

This changes ensure that checks that driver specific checks
are only performed when that driver is enabled.

The complication is that some of these checks require
information to be passed between the source and destination
hosts. For example, when comparing CPU flags.

Change-Id: I7389f0b7f03313d7f04b907f481787dadf0716fd