Example config_drive init script, label the config drive
Configuration with DHCP & cloud-init can be painful.  The config_drive is great,
and it avoids disk injection, but there's no example of how to use it.

So here's a little example init.d script for contrib, and a code patch to make
sure the config drive gets a nice volume label.

Change-Id: I22a1d6a824856ca9651b435d0fe54e348ab107fe