Revert "Fix nova-manage backend_add with sr_uuid"
Fixes bug 988596

Change 77495cfe breaks storage manager, as it makes a
wrong assumption about nova-manage. The sr-uuid must be
generated and provided by the xenapi storage manager.

When the volume service is started, SM volume driver will
create a new SR for everything that does not have an
existing sr_uuid and will populate the db with the result.
It will only introduce the SRs that do have one. The introduce
will work only if the uuid provided is something SM understands.

This reverts commit 77495cfed17f12f225e11fbc10e0f8db516dadf2.

Change-Id: Ifa35484f1fe274f454bba2ac8727d2d56c18fd75